refactor: make technisync a package
